LRSD-4324 Rename
