+++ workspaces/liferay-testray-workspace/modules/testray-rest-api/bnd.bnd
@@ -3,5 +3,5 @@ Bundle-SymbolicName: com.liferay.testray.rest.api
 Bundle-Version: 1.0.0
 Export-Package:\
 	com.liferay.testray.rest.dto.v1_0,\
-	com.liferay.testray.rest.resource.v1_0,\
-	com.liferay.testray.rest.service
+	com.liferay.testray.rest.manager,\
+	com.liferay.testray.rest.resource.v1_0
